requirements.txt
django
djangorestframework
django-celery-beat
django-filter
docker
redis
django-jsonfield
django-html-field
